feat: separate logic to lib and represent a
